Move tile download worker and manager under persistence
