+++ gnd/src/main/java/com/google/android/gnd/repository/OfflineAreaRepository.java
@@ -27,9 +27,9 @@ import com.google.android.gnd.model.basemap.OfflineArea.State;
 import com.google.android.gnd.model.basemap.tile.Tile;
 import com.google.android.gnd.persistence.geojson.GeoJsonParser;
 import com.google.android.gnd.persistence.local.LocalDataStore;
+import com.google.android.gnd.persistence.sync.TileDownloadWorkManager;
 import com.google.android.gnd.persistence.uuid.OfflineUuidGenerator;
 import com.google.android.gnd.ui.util.FileUtil;
-import com.google.android.gnd.workers.TileDownloadWorkManager;
 import com.google.common.collect.ImmutableList;
 import io.reactivex.Completable;
 import io.reactivex.Flowable;
